requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.6; python_version >= '3.6'
aiosignal==1.3.1; python_version >= '3.7'
annotated-types==0.6.0; python_version >= '3.8'
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3; python_version >= '3.7'
attrs==23.1.0; python_version >= '3.7'
automat==22.10.0
blis==0.7.11
catalogue==2.0.10; python_version >= '3.6'
certifi==2023.7.22; python_version >= '3.6'
cffi==1.16.0; python_version >= '3.8'
charset-normalizer==3.3.0; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
cloudpathlib==0.15.1; python_version >= '3.7'
confection==0.1.3; python_version >= '3.6'
constantly==15.1.0
cryptography==41.0.4; python_version >= '3.7'
cssselect==1.2.0; python_version >= '3.7'
cymem==2.0.8
dearpygui==1.10.0; python_version >= '3.7'
filelock==3.12.4; python_version >= '3.8'
frozenlist==1.4.0; python_version >= '3.8'
hydra-core==1.3.2
hyperlink==21.0.0
idna==3.4; python_version >= '3.5'
incremental==22.10.0
iniconfig==2.0.0; python_version >= '3.7'
itemadapter==0.8.0; python_version >= '3.7'
itemloaders==1.1.0; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
jmespath==1.0.1; python_version >= '3.7'
langcodes==3.3.0; python_version >= '3.6'
lxml==4.9.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
markupsafe==2.1.3; python_version >= '3.7'
multidict==6.0.4; python_version >= '3.7'
murmurhash==1.0.10; python_version >= '3.6'
numpy==1.26.0; python_version >= '3.9'
omegaconf==2.3.0; python_version >= '3.6'
openai==0.28.1; python_full_version >= '3.7.1'
packaging==23.2; python_version >= '3.7'
parsel==1.8.1; python_version >= '3.7'
pathy==0.10.2; python_version >= '3.6'
pluggy==1.3.0; python_version >= '3.8'
preshed==3.0.9; python_version >= '3.6'
protego==0.3.0; python_version >= '3.7'
pyasn1==0.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pycparser==2.21
pydantic==2.4.2; python_version >= '3.7'
pydantic-core==2.10.1; python_version >= '3.7'
pydispatcher==2.0.7; platform_python_implementation == 'CPython'
pyfiglet==1.0.2; python_version >= '3.9'
pyopenssl==23.2.0; python_version >= '3.6'
pytest==7.4.2; python_version >= '3.7'
pyyaml==6.0.1; python_version >= '3.6'
queuelib==1.6.2; python_version >= '3.5'
regex==2023.10.3; python_version >= '3.7'
requests==2.31.0; python_version >= '3.7'
requests-file==1.5.1
scrapy==2.11.0; python_version >= '3.8'
service-identity==23.1.0; python_version >= '3.8'
setuptools==68.2.2; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smart-open==6.4.0; python_version >= '3.6' and python_version < '4.0'
spacy==3.7.1; python_version >= '3.7'
spacy-legacy==3.0.12; python_version >= '3.6'
spacy-loggers==1.0.5; python_version >= '3.6'
srsly==2.4.8; python_version >= '3.6'
thinc==8.2.1; python_version >= '3.6'
tiktoken==0.5.1; python_version >= '3.8'
tldextract==3.6.0; python_version >= '3.7'
tqdm==4.66.1; python_version >= '3.7'
twisted==22.10.0; python_full_version >= '3.7.1'
typer==0.9.0; python_version >= '3.6'
typing-extensions==4.8.0; python_version >= '3.8'
urllib3==2.0.6; python_version >= '3.7'
w3lib==2.1.2; python_version >= '3.7'
wasabi==1.1.2; python_version >= '3.6'
weasel==0.3.2; python_version >= '3.6'
yarl==1.9.2; python_version >= '3.7'
zope.interface==6.1; python_version >= '3.7'
